refactor(mcp): extract config.ts + types.ts from McpServer
Phase 1 of McpServer.ts decomposition. Pure move of configuration
constants, the default logger, and all shared interfaces out of the
2400-line monolith. FigmaViewProvider re-exported via the barrel from
./types; no behavior change, no public API change.
