+++ src/main/MCP/index.ts
@@ -1,2 +1,2 @@
 export { McpServer } from "./McpServer";
-export type { FigmaViewProvider } from "./McpServer";
+export type { FigmaViewProvider } from "./types";
